little dev options refactor
