+++ Makefile
@@ -1,2 +1,2 @@
 run_app:
-	streamlit run tsod/active_learning/app/Outlier_Annotation.py
+	streamlit run tsod/active_learning/app/1_Outlier_Annotation.py
